ints to bytes, tele-op should work, timers are off, driving should work, cleaned up some stuff, fixed up some cosmetic issues
NIkita, i dont know what you changed to get auton working. So that is
not fixed, please test.

+++ src/com/frc4343/robot2/RobotTemplate.java
@@ -74,9 +74,9 @@ public class RobotTemplate extends IterativeRobot {
     }
 
     private void climbingHandler() {
-        if (joystickSystem.getJoystick(1).getRawButton(Mappings.EXTEND_CLIMBING_PISTONS)) {
+        if (joystickSystem.getJoystick((byte) 1).getRawButton(Mappings.EXTEND_CLIMBING_PISTONS)) {
             climbingPiston.extend();
-        } else if (joystickSystem.getJoystick(1).getRawButton(Mappings.RETRACT_CLIMBING_PISTONS)) {
+        } else if (joystickSystem.getJoystick((byte) 1).getRawButton(Mappings.RETRACT_CLIMBING_PISTONS)) {
             climbingPiston.retract();
         }
     }
